Refactor: modularise seeding
