+++ packages/common/prisma/seed/seed-constants.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared constants used across the seed orchestration and assembly modules.
+ */
+export const orgName = 'guardian';
+export const cqSourceName = 'seed';
+export const workflowDirectory = '.github/workflows';
+export const defaultWorkflowPath = 'ci.yaml';
